HUM-93 - refactor email verification pages
