+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-email-verification.ts
@@ -1,0 +1,18 @@
+import { useVerifyEmailQuery } from '@/modules/worker/services/email-verification';
+import { getErrorMessageForError } from '@/shared/errors';
+
+export function useEmailVerification(token: string) {
+  const {
+    error: emailVerificationError,
+    isError: isEmailVerificationError,
+    isPending: isEmailVerificationPending,
+  } = useVerifyEmailQuery({ token });
+
+  return {
+    error: emailVerificationError
+      ? getErrorMessageForError(emailVerificationError)
+      : null,
+    isError: isEmailVerificationError,
+    isPending: isEmailVerificationPending,
+  };
+}
