+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-router-state.ts
@@ -1,0 +1,11 @@
+import { useLocationState } from '@/modules/worker/hooks/use-location-state';
+import { routerStateSchema } from '@/modules/worker/types/email-verification.types';
+
+export function useRouterState() {
+  const { field: routerState } = useLocationState({
+    keyInStorage: 'routerState',
+    schema: routerStateSchema,
+  });
+
+  return routerState;
+}
