+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-verification-token.ts
@@ -1,0 +1,14 @@
+import { useLocationState } from '@/modules/worker/hooks/use-location-state';
+import { tokenSchema } from '@/modules/worker/types/email-verification.types';
+
+export function useVerificationToken() {
+  const { field: token } = useLocationState({
+    schema: tokenSchema,
+    locationStorage: 'search',
+  });
+
+  return {
+    token,
+    isLoading: token === undefined,
+  };
+}
