+++ packages/apps/human-app/frontend/src/modules/worker/views/email-verification/worker-email-verification-process.page.tsx
@@ -1,0 +1,13 @@
+import { EmailVerificationProcess } from '@/modules/worker/components/email-verification/email-verification-process';
+import { useVerificationToken } from '@/modules/worker/hooks/use-verification-token';
+import { PageCardLoader } from '@/shared/components/ui/page-card';
+
+export function WorkerEmailVerificationProcessPage() {
+  const { token, isLoading } = useVerificationToken();
+
+  if (isLoading || !token) {
+    return <PageCardLoader />;
+  }
+
+  return <EmailVerificationProcess token={token} />;
+}
