+++ packages/apps/human-app/frontend/src/modules/worker/views/email-verification/worker-verify-email.page.tsx
@@ -1,0 +1,5 @@
+import { EmailVerificationFormContainer } from '@/modules/worker/components/email-verification/email-verification-form-container';
+
+export function WorkerVerifyEmailPage() {
+  return <EmailVerificationFormContainer />;
+}
