+++ packages/apps/human-app/frontend/src/router/routes.tsx
@@ -17,8 +17,8 @@ import type { PageHeaderProps } from '@/shared/components/layout/protected/page-
 import { SetUpOperatorPage } from '@/modules/operator/views/sign-up/set-up-operator.page';
 import { EditExistingKeysSuccessPage } from '@/modules/operator/views/sign-up/edit-existing-keys-success.page';
 import { AddKeysOperatorPage } from '@/modules/operator/views/sign-up/add-keys.page';
-import { VerifyEmailWorkerPage } from '@/modules/worker/views/email-verification/verify-email.page';
-import { EmailVerificationWorkerPage } from '@/modules/worker/views/email-verification/email-verification.page';
+import { WorkerVerifyEmailPage } from '@/modules/worker/views/email-verification/worker-verify-email.page';
+import { WorkerEmailVerificationProcessPage } from '@/modules/worker/views/email-verification/worker-email-verification-process.page';
 import { AddStakeOperatorPage } from '@/modules/operator/views/sign-up/add-stake.page';
 import { ConnectWalletOperatorPage } from '@/modules/operator/views/sign-up/connect-wallet.page';
 import { OperatorProfilePage } from '@/modules/operator/views/profile/profile.page';
@@ -57,11 +57,11 @@ export const unprotectedRoutes: RouteProps[] = [
   },
   {
     path: routerPaths.worker.emailVerification,
-    element: <EmailVerificationWorkerPage />,
+    element: <WorkerEmailVerificationProcessPage />,
   },
   {
     path: routerPaths.worker.verifyEmail,
-    element: <VerifyEmailWorkerPage />,
+    element: <WorkerVerifyEmailPage />,
   },
   {
     path: routerPaths.worker.sendResetLink,
